Added support for displaying errata within the DRM and placed the errata about the incorrect ordering of \cleanup and \exception clauses in the \block macro section. I also updated the original errata text to match the in-situ text.
Going forward, we'll place semantically significant errata within the DRM
content.

Errata are paragraphs with class "errata" and should start with the text
"<em>Errata:</em> ", which is dynamically replaced by the CSS. Errata are
displayed inline by default, and can be placed within the left margin or
floating within the right of the document body by adding the "left" or
"right" classes, respectively. Errata displayed within the left margin have
much less room to work with, but keep out of the way of document
content. Errata displayed on the right are wider, but can only be used
where there is sufficient blank space in the content. Inline errata should
only be used when there is not sufficient space to the left or right, as it
interrupts the book content.