A new, cleaner design for the online DRM that more closely matches the visual style of the printed DRM.
The width of the navigation area is now a percentage of the viewport
width, allowing users to resize it (along with the content area) by
changing their browser window width.

The book title in the navigation area is now a link to the book start,
and it is an image that scales with the font size and viewport width.

Lots of XHTML/CSS cleanups and improvements for better rendering in
more browsers, including the Sony PSP.

Added an alternate style with a horizontal navigation bar across the
top. It is the default stylesheet for handheld devices. It is still a
work in progress, though, and doesn't entirely render as desired on
Firefox, although it looks good in Safari and the Sony PSP browser.